--- apps/web/src/components/Shared/Account/Fund/TransferFund/index.tsx
@@ -1,58 +1,0 @@
-import Loader from "@/components/Shared/Loader";
-import { Image } from "@/components/Shared/UI";
-import getTokenImage from "@/helpers/getTokenImage";
-import { useFundModalStore } from "@/store/non-persisted/modal/useFundModalStore";
-import { useAccountStore } from "@/store/persisted/useAccountStore";
-import { NATIVE_TOKEN_SYMBOL } from "@hey/data/constants";
-import { useAccountBalancesQuery } from "@hey/indexer";
-import Transfer from "./Transfer";
-
-const TransferFund = () => {
-  const { currentAccount } = useAccountStore();
-  const { token } = useFundModalStore();
-  const { data: balance, loading } = useAccountBalancesQuery({
-    variables: {
-      request: {
-        ...(token
-          ? { tokens: [token?.contractAddress] }
-          : { includeNative: true })
-      }
-    },
-    pollInterval: 3000,
-    skip: !currentAccount?.address,
-    fetchPolicy: "no-cache"
-  });
-
-  if (loading) {
-    return <Loader message="Loading balance..." className="my-10" />;
-  }
-
-  const tokenBalance =
-    balance?.accountBalances[0].__typename === "Erc20Amount"
-      ? Number(balance.accountBalances[0].value).toFixed(2)
-      : balance?.accountBalances[0].__typename === "NativeAmount"
-        ? Number(balance.accountBalances[0].value).toFixed(2)
-        : 0;
-
-  return (
-    <div className="m-5">
-      <div className="flex flex-col items-center gap-2 text-center">
-        <Image
-          className="size-12 rounded-full"
-          src={getTokenImage(token?.symbol)}
-          alt={token?.symbol}
-        />
-        <div className="font-bold text-2xl">
-          {tokenBalance} {token?.symbol ?? NATIVE_TOKEN_SYMBOL}
-        </div>
-        <div className="text-gray-500 text-sm dark:text-gray-200">
-          {token?.symbol ?? NATIVE_TOKEN_SYMBOL} enables you to do various
-          Hey-specific actions
-        </div>
-      </div>
-      <Transfer token={token} />
-    </div>
-  );
-};
-
-export default TransferFund;
