+++ apps/web/src/components/Shared/Modal/Subscribe.tsx
@@ -16,8 +16,8 @@ import {
 import { useState } from "react";
 import { erc20Abi, formatUnits } from "viem";
 import { useReadContract } from "wagmi";
-import SwapButton from "../Account/Fund/SwapButton";
 import Loader from "../Loader";
+import SwapButton from "../SwapButton";
 
 const Subscribe = () => {
   const { currentAccount } = useAccountStore();
