--- packages/indexer/documents/mutations/account/funds/Deposit.graphql
@@ -1,16 +1,0 @@
-mutation Deposit($request: DepositRequest!) {
-  deposit(request: $request) {
-    ... on SelfFundedTransactionRequest {
-      ...SelfFundedTransactionRequest
-    }
-    ... on SponsoredTransactionRequest {
-      ...SponsoredTransactionRequest
-    }
-    ... on TransactionWillFail {
-      ...TransactionWillFail
-    }
-    ... on InsufficientFunds {
-      reason
-    }
-  }
-}
